Revert "Inital refactor commit"
