+++ WeatherApplication/WeatherApplication.csproj
@@ -64,7 +64,6 @@
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </ApplicationDefinition>
-    <Compile Include="StoreAPIData.cs" />
     <Page Include="Forecast.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
@@ -84,7 +83,6 @@
     </Compile>
     <Compile Include="GetAPI.cs" />
     <Compile Include="GetDarkSky.cs" />
-    <Compile Include="GetForecast.cs" />
     <Compile Include="GetWeather.cs" />
     <Compile Include="MainWindow.xaml.cs">
       <DependentUpon>MainWindow.xaml</DependentUpon>
